@@ -42,7 +42,7 @@ class _OtpWidgetState extends BaseState<OtpWidget> {
                   textAlign: TextAlign.center,
                   style:  TextStyle(
                     fontWeight: FontWeight.w500,
-                    color: context.textTheme().headline6?.color,
+                    color: context.textTheme().titleLarge?.color,
                     fontSize: 15.0,
                   ),
                 ),
